+++ src/spell/modules/spell/spell/src/script.rs
@@ -101,8 +101,8 @@ pub fn script_cid() -> CIDValue {
 #[test_env_helpers::after_each]
 #[cfg(test)]
 mod tests {
-    use marine_rs_sdk_test::CallParameters;
     use marine_rs_sdk_test::marine_test;
+    use marine_rs_sdk_test::CallParameters;
 
     use crate::schema::DB_FILE;
 
@@ -147,7 +147,10 @@ mod tests {
             second_set.success,
             "set_script_source_to_file returned false (fail), expected true (success)"
         );
-        assert_eq!(spell.get_script_source_from_file().source_code, "(seq (null) (null))");
+        assert_eq!(
+            spell.get_script_source_from_file().source_code,
+            "(seq (null) (null))"
+        );
     }
 
     #[marine_test(